[Fix] 팔로우 기능 보완 #16
스스로 팔로우 불가, 중복 팔로우 방지

+++ src/main/java/verbly/spring/domain/follow/service/FollowService.java
@@ -1,9 +1,11 @@
 package verbly.spring.domain.follow.service;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import verbly.spring.domain.follow.entity.Follow;
+import verbly.spring.domain.follow.exception.FollowHandler;
 import verbly.spring.domain.follow.repo.FollowRepository;
 import verbly.spring.domain.user.dto.response.UserResponseDTO;
 import verbly.spring.domain.user.entity.User;
@@ -21,7 +23,15 @@ public class FollowService {
     private final FollowRepository followRepository;
     private final UserRepository userRepository;
 
+    @Transactional
     public void createFollowing(Long followerId, Long followeeId) {
+
+        if(followerId.equals(followeeId))
+            throw new FollowHandler(ErrorStatus.CANT_SELF_FOLLOW);
+
+        if(isFollowed(followerId, followeeId)) {
+            throw new FollowHandler(ErrorStatus.ALREADY_FOLLOWED);
+        }
 
         Optional<User> optionalFollower = userRepository.findById(followerId);
         if(optionalFollower.isEmpty()) {
@@ -39,6 +49,7 @@ public class FollowService {
         followRepository.save(follow);
     }
 
+    @Transactional
     public List<UserResponseDTO.FollowRecommendUserResponseDTO> getRecommendFollowList(Long followerId) {
         List<User> followRecommendUserList = followRepository.findRandomUser(followerId, PageRequest.of(0, 3));
 
@@ -48,8 +59,17 @@ public class FollowService {
                 .toList();
     }
 
+    @Transactional
     public void unfollow(Long followerId, Long followeeId) {
 
+        if(!isFollowed(followerId, followeeId))
+            throw new FollowHandler(ErrorStatus.NOT_FOLLOWED_USER);
+
         followRepository.deleteByFollowerIdAndFolloweeId(followerId, followeeId);
+    }
+
+    public boolean isFollowed(Long followerId, Long followeeId) {
+
+        return followRepository.existsFollowByFollowerIdAndFolloweeId(followerId, followeeId);
     }
 }
